% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy3D.R
\name{energy3D}
\alias{energy3D}
\title{Energy Function for Trivariate Spline over Tetrahedral Partition}
\usage{
energy3D(V, Tr, d)
}
\arguments{
\item{V}{The \code{N} by three matrix of vertices of a tetrahedron, where \code{N} is the number of vertices. Each row is the coordinates for a vertex.
\cr}

\item{d}{The degree of piecewise polynomials -- default is 9, and usually \code{d} is greater than one.
\cr}

\item{Th}{The tetrahedral partition matrix of dimension \code{nT} by four, where \code{nT} is the number of tetrahedrons in the partition. Each row is the indices of vertices in \code{V}.
\cr}
}
\value{
The energy matrix.
}
\description{
Energy Function for Trivariate Spline over Tetrahedral Partition
}
\details{
This R program is modified based on the Matlab program written by Ming-Jun Lai from the University of Georgia.
}
\examples{
# example 1
d <- 4
V <- matrix(rbind(c(0, 0, 0), c(0, 1, 0), c(1, 0, 0), c(1, 1, 0),
c(0, 0, 1), c(0, 1, 1), c(1, 0, 1), c(1, 1, 1)), ncol = 3)
Tr <- matrix(rbind(c(5, 1, 2, 3), c(6, 5, 2, 3), c(6, 7, 5, 3),
c(6, 4, 7, 3), c(6, 2, 4, 3), c(6, 8, 7, 4)), ncol = 4)
P <- energy3D(V, Tr, d)
}
