% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.MPST.R
\name{fit.mpst.d}
\alias{fit.mpst.d}
\title{Fit an MPST Model Using Distributed Learning}
\usage{
\method{fit}{mpst.d}(ind.Tr, Y, Z, V, Tr, d = NULL, r = 1, lambda, nl, load.all)
}
\arguments{
\item{ind.Tr}{Index of the sub-triangulation.}

\item{Y}{The response variable vector.}

\item{Z}{A matrix of observation coordinates.}

\item{V}{A matrix of vertices of the triangulation.}

\item{Tr}{A triangulation matrix.}

\item{d}{The degree of piecewise polynomials.}

\item{r}{The smoothness parameter.}

\item{lambda}{The tuning parameter.}

\item{nl}{Number of lambda values to consider.}

\item{load.all}{Preloaded data for distributed processing.}
}
\value{
A list containing distributed model fit components.
}
\description{
Internal function to fit a multivariate penalized spline over 
triangulation (MPST) using distributed learning.
}
\keyword{internal}
