% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.MPST.R
\name{fit.mpst.g}
\alias{fit.mpst.g}
\title{Fit an MPST Model Using Global Learning}
\usage{
\method{fit}{mpst.g}(Y, Z, V, Tr, d = NULL, r = 1, lambda)
}
\arguments{
\item{Y}{The response variable vector.}

\item{Z}{A matrix of observation coordinates.}

\item{V}{A matrix of vertices of the triangulation.}

\item{Tr}{A triangulation matrix.}

\item{d}{The degree of piecewise polynomials.}

\item{r}{The smoothness parameter.}

\item{lambda}{The tuning parameter.}
}
\value{
A list containing global model fit components.
}
\description{
Internal function to fit a multivariate penalized spline over 
triangulation (MPST) using global learning.
}
\keyword{internal}
