% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.MPST.R
\name{fit.mpst.internal}
\alias{fit.mpst.internal}
\title{Internal Function for MPST Model Fitting}
\usage{
\method{fit}{mpst.internal}(Y, Z, V, Tr, d = NULL, r = 1, lambda, nl = 1, method, P.func)
}
\arguments{
\item{Y}{The response variable vector.}

\item{Z}{A matrix of observation coordinates.}

\item{V}{A matrix of vertices of the triangulation.}

\item{Tr}{A triangulation matrix.}

\item{d}{The degree of piecewise polynomials.}

\item{r}{The smoothness parameter.}

\item{lambda}{The tuning parameter.}

\item{nl}{Number of lambda values to consider.}

\item{method}{Learning method: `"G"` for global or `"D"` for distributed.}

\item{P.func}{Parallelization method for distributed learning.}
}
\value{
A list containing model fit components.
}
\description{
Internal function to fit a multivariate penalized spline over triangulation 
(MPST) using global or distributed learning.
}
\keyword{internal}
