% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPST.R
\name{fit}
\alias{fit}
\title{Fit a Model}
\usage{
fit(object, ...)
}
\arguments{
\item{object}{An object to be fitted.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
Depends on the specific method used.
}
\description{
A generic function for fitting models.
}
