% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.MPST.R
\name{plot.contour.mpst}
\alias{plot.contour.mpst}
\title{Plot Contour for MPST Models}
\usage{
\method{plot}{contour.mpst}(mfit, Zgrid = NULL)
}
\arguments{
\item{mfit}{An MPST model fit object.}

\item{Zgrid}{An optional grid for plotting. If NULL, a grid will be generated.}
}
\value{
A plotly object representing the contour plot.
}
\description{
Generates a contour plot for a 3D MPST model. This function is used internally
by `plot.MPST` and is not intended for direct use by end users.
}
\keyword{internal}
