% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.MPST.R
\name{plot.surface.mpst}
\alias{plot.surface.mpst}
\title{Plot Surface for MPST Models}
\usage{
\method{plot}{surface.mpst}(mfit, Zgrid = NULL)
}
\arguments{
\item{mfit}{An MPST model fit object.}

\item{Zgrid}{An optional grid for plotting. If NULL, a grid is generated automatically.}
}
\value{
A plotly object representing the surface plot.
}
\description{
Generates a 3D surface plot for an MPST model. This function is used internally
by `plot.MPST` and is not intended for direct use by end users.
}
\keyword{internal}
