% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.MPST.R
\name{pred.mpst}
\alias{pred.mpst}
\title{Predict Values for New Data}
\usage{
pred.mpst(mfit, Znew = NULL)
}
\arguments{
\item{mfit}{A fitted MPST model object.}

\item{Znew}{A matrix of new coordinates for prediction.}
}
\value{
A list containing:
- `Ypred`: Predicted values.
- `ind.inside`: Indices of the points inside the domain.
- `d`: Degree of the spline.
- `N.cores`: Number of cores used for computation.
}
\description{
Internal function for making predictions using a fitted MPST model.
}
\keyword{internal}
