% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.MPST.R
\name{predict.MPST}
\alias{predict.MPST}
\title{Make predictions from a fitted BPST object.}
\usage{
\method{predict}{MPST}(mfit, Znew = NULL)
}
\arguments{
\item{mfit}{Fitted ``MPST" object.
\cr}

\item{Znew}{The cooridinates of new locationsfor prediction -- default is the observed coordinates, \code{Z}.
\cr}
}
\value{
A vector of predicted values is returned.
}
\description{
This function is used to make predictions of a fitted BPST object.
}
\details{
This R program is modified based on the Matlab program written by Ming-Jun Lai from the University of Georgia and Li Wang from the Iowa State University.
}
\examples{
data(VT.square) 
d = 5; r = 1; 
func = 1; sigma = 1;
n = 2000;
Z = matrix(runif(2*n, 0, 1), nrow = n, ncol = 2)
sam = dataGenerator2D(Z, V, Tr, func, sigma)
Y = as.vector(sam$Y); Z = as.matrix(sam$Z);
mfit = fit.MPST(Y, Z, V, Tr, d, r)
# Grid points for evaluation
n1.grid = 51; n2.grid = 51; n.grid = n1.grid * n2.grid;
u.grid = seq(0, 1, length.out = n1.grid)
v.grid = seq(0, 1, length.out = n2.grid)
uu.grid = rep(u.grid, each = n2.grid)
vv.grid = rep(v.grid, times = n1.grid)
Z.grid = as.matrix(cbind(uu.grid, vv.grid))
pop = dataGenerator2D(Z.grid, V, Tr, func, sigma)
Y.grid = pop$Y; mu.grid = pop$mu;
mpred = predict.MPST(mfit, Z.grid)
rmspe = sqrt(mean((Y.grid - mpred$Ypred)^2, na.rm = TRUE)); rmspe

}
