% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPST.R
\name{predict}
\alias{predict}
\title{Predict from a Model}
\usage{
predict(object, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
Depends on the specific method used.
}
\description{
A generic function for making predictions from models.
}
