% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataGenerator2D.R
\name{dataGenerator2D}
\alias{dataGenerator2D}
\title{Generate testing dataset for bivariate spline smoothing.}
\usage{
dataGenerator2D(Z, V, Tr, func, sigma, seed)
}
\arguments{
\item{Z}{The cooridinates of dimension \code{n} by two. Each row is the coordinates of a point.
\cr}

\item{V}{The \code{nV} by two matrix of vertices of a triangulation, where \code{nV} is the number of vertices. Each row is the coordinates for a vertex.
\cr}

\item{Tr}{The triangulation matrix of dimention \code{nT} by three, where \code{nT} is the number of triangles in the triangulation. Each row is the indices of vertices in \code{V}.
\cr}

\item{func}{The choice of test function -- default is 1. Possible choices include 1, 2, 3, 4, 5, 6, 7, 8.
\cr}

\item{sigma}{The standard deviation of the white noise --  default is 0.1.
\cr}
}
\value{
A list of vectors and matrice, including:
\item{Y}{The response variable.}
\item{mu}{The mean function.}
\item{Z}{The coordinates.}
\item{ind.inside}{A vector contains the indicators whether the point is inside the given triangulation.}
\item{ind.T}{A vector contains the indexes of which triangle the points falls in.}
}
\description{
This function generate the testing dataset for bivariate spline smoothing.
}
\details{
This R program is modified based on Lai and Wang (2013).
}
\keyword{internal}
