% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPST.R
\name{m}
\alias{m}
\title{Generate MPST Parameters}
\usage{
m(..., V = NULL, Tr = NULL, d, r)
}
\arguments{
\item{...}{Additional arguments (not used).}

\item{V}{Matrix of vertices (optional).}

\item{Tr}{Matrix of triangulations (optional).}

\item{d}{Degree of the spline (must be an integer ≥ 1).}

\item{r}{Smoothness parameter (must be an integer).}
}
\value{
A list containing parameters `d` and `r`.
}
\description{
Internal function to create a list of parameters for MPST.
}
\keyword{internal}
