% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.MPST.R
\name{plot.MPST}
\alias{plot.MPST}
\title{Plot Different Views for MPST Models}
\usage{
\method{plot}{MPST}(x, Zgrid = NULL, mview = NULL, ...)
}
\arguments{
\item{x}{An MPST model fit object.}

\item{Zgrid}{An optional grid for plotting. If NULL, it will be generated automatically.}

\item{mview}{A character string specifying the view to plot. Must be one of 'contour', 
'surface', or 'slice'.}

\item{...}{Additional arguments passed to the specific plot functions.}
}
\value{
No return value. The function generates plots.
}
\description{
This function allows plotting different views (contour, surface, or slice) 
for an MPST model fit object.
}
\examples{
\dontrun{
plot.MPST(model_fit_object, mview = "contour")
}
}
