% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.MPST.R
\name{plot.slice.mpst}
\alias{plot.slice.mpst}
\title{Plot Slices for 3D MPST Models}
\usage{
\method{plot}{slice.mpst}(mfit, Zgrid = NULL)
}
\arguments{
\item{mfit}{An MPST model fit object.}

\item{Zgrid}{An optional grid for plotting. If NULL, a grid is generated automatically.}
}
\value{
A manipulate object that provides interactive slices through the 3D array.
}
\description{
Generates interactive 2D slices for a 3D MPST model. This function is used internally
by `plot.MPST` and is not intended for direct use by end users.
}
\keyword{internal}
