% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPST.R
\name{print.MPST}
\alias{print.MPST}
\title{Print Method for MPST Models}
\usage{
\method{print}{MPST}(x, ...)
}
\arguments{
\item{x}{An MPST object.}
}
\value{
No return value; prints to the console.
}
\description{
Prints the summary of an MPST model.
}
