% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPST.R
\name{print.MPST}
\alias{print.MPST}
\title{Print Method for MPST Models}
\usage{
\method{print}{MPST}(x, digits = 4, max = 10)
}
\arguments{
\item{x}{An MPST object.}

\item{digits}{Number of digits to round numeric outputs. Default is 4.}

\item{max}{Number of rows to display for long outputs. Default is 10.}
}
\value{
No return value; prints to the console.
}
\description{
Prints the summary of an MPST model.
}
